Added Chinese Remainder Theorem
